tests/unittests/test-is_mailaddress.cpp 以外の部分に関して a722f9c86f2860606b3f9e7dc79bb169294cfaea を revert する前の状態にする tests/unittests/test-is_mailaddress.cpp で ASSERT_SAME と ASSERT_CHANGE の定義で IsMailAddress のみ確認するようにする _NEW_IMPL および _OLD_IMPL を削除する
