+++ sakura_core/view/colors/CColor_Url.cpp
@@ -6,7 +6,6 @@
 #include "doc/CEditDoc.h"
 #include "doc/layout/CLayout.h"
 #include "types/CTypeSupport.h"
-#include "util/string_ex.h" // IsURL
 
 // -- -- -- -- -- -- -- -- -- -- -- -- -- -- -- -- -- -- -- -- //
 //                           URL                               //
@@ -19,7 +18,7 @@ bool CColor_Url::BeginColor(const CStringRef& cStr, int nPos)
 	int	nUrlLen;
 
 	if( _IsPosKeywordHead(cStr,nPos) /* URLを表示する */
-	 && IsURL( cStr.GetPtr() + nPos, cStr.GetLength() - nPos, &nUrlLen )	/* 指定アドレスがURLの先頭ならばTRUEとその長さを返す */
+	 && IsURL( cStr.GetPtr(), nPos, cStr.GetLength(), &nUrlLen )	/* 指定アドレスがURLの先頭ならばTRUEとその長さを返す */
 	){
 		this->m_nCOMMENTEND = nPos + nUrlLen;
 		return true;
